modularize time-picker, input, input-number,switch, select, radio, rate, slider, tree-select, mention, transfer
