+++ src/index.ts
@@ -1,21 +1,21 @@
 export * from "./date-picker";
-export * from "./TimePicker";
+export * from "./time-picker";
 export * from "./FormikDebug";
-export * from "./Input";
-export * from "./InputNumber";
-export * from "./Switch";
+export * from "./input";
+export * from "./input-number";
+export * from "./switch";
 export * from "./FormItem";
-export * from "./Select";
+export * from "./select";
 export * from "./SubmitButton";
 export * from "./ResetButton";
 export * from "./HtmlText";
 export * from "./checkbox";
-export * from "./Radio";
+export * from "./radio";
 export * from "./auto-complete";
-export * from "./Rate";
-export * from "./Slider";
+export * from "./rate";
+export * from "./slider";
 export * from "./cascader";
 export * from "./Form";
-export * from "./TreeSelect";
-export * from "./Mention";
-export * from "./Transfer";
+export * from "./tree-select";
+export * from "./mention";
+export * from "./transfer";
