+++ src/input-number/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/input-number/style/css.js"
