+++ src/input/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/input/style/css.js"
