+++ src/mention/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/mention/style/css.js"
