+++ src/rate/index.tsx
@@ -1,7 +1,7 @@
 import { Rate as $Rate } from "antd";
 import { Field, FieldProps } from "formik";
 import * as React from "react";
-import { FormikFieldProps } from "./FieldProps";
+import { FormikFieldProps } from "../FieldProps";
 import { RateProps as $RateProps } from "antd/lib/rate";
 
 export type RateProps = FormikFieldProps & $RateProps;
@@ -20,3 +20,5 @@ export const Rate = ({ name, validate, ...restProps }: RateProps) => (
     )}
   </Field>
 );
+
+export default Rate
