+++ src/rate/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/rate/style/css.js"
