+++ src/select/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/select/style/css.js"
