+++ src/select/style/index.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/select/style"
