+++ src/slider/index.tsx
@@ -1,7 +1,7 @@
 import { Slider as $Slider } from "antd";
 import { Field, FieldProps } from "formik";
 import * as React from "react";
-import { FormikFieldProps } from "./FieldProps";
+import { FormikFieldProps } from "../FieldProps";
 import { SliderProps as $SliderProps } from "antd/lib/slider";
 
 export type SliderProps = FormikFieldProps & $SliderProps;
@@ -20,3 +20,5 @@ export const Slider = ({ name, validate, ...restProps }: SliderProps) => (
     )}
   </Field>
 );
+
+export default Slider
