+++ src/switch/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/switch/style/css.js"
