+++ src/time-picker/index.tsx
@@ -2,7 +2,7 @@ import { TimePicker as $TimePicker } from "antd";
 import { Field, FieldProps } from "formik";
 import * as React from "react";
 import moment from "moment";
-import { FormikFieldProps } from "./FieldProps";
+import { FormikFieldProps } from "../FieldProps";
 import { TimePickerProps as $TimePickerProps } from "antd/lib/time-picker";
 
 export type TimePickerProps = FormikFieldProps & $TimePickerProps;
@@ -21,3 +21,5 @@ export const TimePicker = ({ name, validate, ...restProps }: TimePickerProps) =>
     )}
   </Field>
 );
+
+export default TimePicker
