+++ src/time-picker/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/time-picker/style/css.js"
