+++ src/time-picker/style/index.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/time-picker/style"
