+++ src/tree-select/index.tsx
@@ -1,7 +1,7 @@
 import { TreeSelect as $TreeSelect } from "antd";
 import { Field, FieldProps } from "formik";
 import * as React from "react";
-import { FormikFieldProps } from "./FieldProps";
+import { FormikFieldProps } from "../FieldProps";
 import { TreeSelectProps as $TreeSelectProps } from "antd/lib/tree-select";
 
 export type TreeSelectProps = FormikFieldProps & $TreeSelectProps & { children: React.ReactNode };
@@ -22,5 +22,6 @@ export const TreeSelect = ({ name, validate, onChange, ...restProps }: TreeSelec
   </Field>
 );
 
+export default TreeSelect
 
 TreeSelect.TreeNode = $TreeSelect.TreeNode;
