+++ src/tree-select/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/tree-select/style/css.js"
